refactor: getFromCookies
