+++ src/components/Providers.tsx
@@ -6,7 +6,7 @@ import { ParticleProvider } from '@/components/ParticleProvider.js';
 import { QueryClientProviders } from '@/components/QueryClientProviders.js';
 import { SolanaWalletAdapterProvider } from '@/components/SolanaWalletAdapterProvider.js';
 import { WagmiProvider } from '@/components/WagmiProvider.js';
-import { getLocaleFromCookies } from '@/helpers/getLocaleFromCookies.js';
+import { getLocaleFromCookies } from '@/helpers/getFromCookies.js';
 import { getLocale, setupLocaleForSSR } from '@/i18n/index.js';
 
 type ProviderProps = PropsWithChildren<{
