+++ src/store/useLocale.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 
-import { getLocaleFromCookies } from '@/helpers/getLocaleFromCookies.js';
+import { getLocaleFromCookies } from '@/helpers/getFromCookies.js';
 
 export function useLocale() {
     return useMemo(() => getLocaleFromCookies(), []);
